Remove user from cache when removing from DB
When a user is added because they don't exist and then removed, they
should also be removed from the cache dictionary. If they are not, the
next check will return a non-existent userid from the cache and DB
operations will fail, leaving an aborted transaction until the service
is restarted.